508 expanding bar query (#537)
* updated a watch method to constantly watch PAD status

* modified query to include more emails that manage less than 50 corps

* modified test and add more unittests

* fixed linter

* fixed linter

+++ jobs/business-sync/src/business_sync/config.py
@@ -19,6 +19,7 @@ import tempfile
 import atexit
 import base64
 import json
+import logging
 
 from dotenv import find_dotenv, load_dotenv
 
@@ -101,15 +102,15 @@ class _Config:  # pylint: disable=too-few-public-methods
         except (base64.binascii.Error, UnicodeDecodeError, json.JSONDecodeError) as e:
             raise ValueError("Invalid base64-encoded credentials") from e
 
-        # Register cleanup function to delete the temp file at exit
-        @staticmethod
-        def remove_temp_file():
-            """Remove temp file generated"""
-            try:
-                if _Config.WAREHOUSE_CREDENTIALS_FILE_PATH:
-                    os.remove(_Config.WAREHOUSE_CREDENTIALS_FILE_PATH)
-            except OSError:
-                pass
+    @staticmethod
+    def remove_temp_file():
+        """Remove temp file generated"""
+        try:
+            if _Config.WAREHOUSE_CREDENTIALS_FILE_PATH \
+                    and os.path.exists(_Config.WAREHOUSE_CREDENTIALS_FILE_PATH):
+                os.remove(_Config.WAREHOUSE_CREDENTIALS_FILE_PATH)
+        except OSError as err:
+            logging.error("Error removing temp file %s: %s", _Config.WAREHOUSE_CREDENTIALS_FILE_PATH, err)
 
     # Update the config to use the temporary file path
     WAREHOUSE_CREDENTIALS_FILE = WAREHOUSE_CREDENTIALS_FILE_PATH
